getKeysByValue moved
