--- src/utils/get-key-by-value.js
@@ -1,3 +1,0 @@
-export const getKeyByValue = (object, value) => {
-  return Object.keys(object).find((key) => object[key] === value);
-};
