--- src/utils/get-keys-by-value.js
@@ -1,5 +1,0 @@
-export const getKeysByValue = (object, value) => {
-  return Object.keys(object)
-    .filter((key) => object[key] === value)
-    .join(',');
-};
