+++ src/utils/object-utils.test.js
@@ -1,6 +1,8 @@
 import {
   removeChildFromObject,
   removeElementsFromObjectValues,
+  getKeyByValue,
+  getKeysByValue,
 } from './object-utils';
 import { data } from '../components/experiment-tracking/mock-data';
 
@@ -47,4 +49,36 @@ test('return expected runs which should not have the first 3 values', () => {
   );
 
   expect(received).toStrictEqual(expected);
+});
+
+test('return the correct key for the value', () => {
+  const mockObject = {
+    key1: 'value1',
+    key2: 'value2',
+    key3: 'value3',
+    key4: 'value4',
+  };
+
+  const mockValue = 'value3';
+
+  const expected = 'key3';
+  const result = getKeyByValue(mockObject, mockValue);
+
+  expect(result).toEqual(expected);
+});
+
+test('return the correct keys for the value', () => {
+  const mockObject = {
+    key1: 'value1',
+    key2: 'value1',
+    key3: 'value2',
+    key4: 'value3',
+  };
+
+  const mockValue = 'value1';
+
+  const expected = 'key1,key2';
+  const result = getKeysByValue(mockObject, mockValue);
+
+  expect(result).toEqual(expected);
 });
